requirements.txt
requests
sseclient-py
python-dotenv
fastapi==0.112.2
uvicorn[standard]==0.30.6
websockets>=11
